+++ crates/handler/src/pre_execution.rs
@@ -67,6 +67,21 @@ pub fn load_accounts<
 
 /// Validates caller account nonce and code according to EIP-3607.
 #[inline]
+pub fn validate_account_nonce_and_code_with_components(
+    caller_info: &AccountInfo,
+    tx: impl Transaction,
+    cfg: impl Cfg,
+) -> Result<(), InvalidTransaction> {
+    validate_account_nonce_and_code(
+        caller_info,
+        tx.nonce(),
+        cfg.is_eip3607_disabled(),
+        cfg.is_nonce_check_disabled(),
+    )
+}
+
+/// Validates caller account nonce and code according to EIP-3607.
+#[inline]
 pub fn validate_account_nonce_and_code(
     caller_info: &AccountInfo,
     tx_nonce: u64,
@@ -153,12 +168,7 @@ pub fn validate_against_state_and_deduct_caller<
     // Load caller's account.
     let mut caller = journal.load_account_code_mut(tx.caller())?.data;
 
-    validate_account_nonce_and_code(
-        &caller.info,
-        tx.nonce(),
-        cfg.is_eip3607_disabled(),
-        cfg.is_nonce_check_disabled(),
-    )?;
+    validate_account_nonce_and_code_with_components(&caller.info, tx, cfg)?;
 
     let new_balance = calculate_caller_fee(*caller.balance(), tx, block, cfg)?;